introduced capped clustering with ignoring the outliers

+++ src/main/kotlin/solver/model/Result.kt
@@ -4,7 +4,7 @@ data class Result(
     val name: String,
     val size: Int,
     val finalPath: List<Node>,
-    val budget: Double,
+    val budget: Int,
     val successful: Boolean,
     val revenue: Int,
     val budgetSpent: Double,
